+++ NEWS.md
@@ -1,6 +1,7 @@
 #  Changes in version 1.2-4.1 (xx/xx/22)
 
 * changes for the new version of stream.
+* Added DSClassifier_MOA to interface RMOA.
 
 #  Changes in version 1.2-4 (05/09/22)